feat(packages): 🏗  Split package & dependencies related snippets
