+++ .changeset/orange-rats-ring.md
@@ -1,0 +1,6 @@
+---
+"@terminal-nerds/snippets-package": minor
+"@terminal-nerds/snippets-runtime": minor
+---
+
+✨ Split snippets related to package and dependencies into a separate package - `package`
