+++ packages/package/package.json
@@ -9,12 +9,12 @@
 		"typescript",
 		"package"
 	],
-	"license": "MIT",
 	"author": {
 		"name": "Mateusz Kadlubowski",
 		"email": "xeho91@pm.me",
 		"url": "https://xeho91.com"
 	},
+	"license": "MIT",
 	"homepage": "https://github.com/terminal-nerds/snippets",
 	"repository": {
 		"type": "git",
@@ -40,14 +40,30 @@
 	"files": [
 		"dist/"
 	],
+	"dependencies": {
+		"@terminal-nerds/snippets-runtime": "workspace:*",
+		"read-pkg-up": "9.1.0",
+		"type-fest": "3.7.2"
+	},
+	"peerDependencies": {
+		"typescript": "5.0.2"
+	},
+	"peerDependenciesMeta": {
+		"typescript": {
+			"optional": true
+		}
+	},
+	"devDependencies": {
+		"@terminal-nerds/snippets-test": "workspace:*"
+	},
 	"scripts": {
 		"build": "tsup",
 		"clean": "concurrently \"pnpm:clean:*\" --group --timings",
 		"clean:build": "del \"./dist\"",
 		"clean:cache": "del \"./node_modules/.cache\"",
 		"clean:test": "del \"./coverage\"",
-		"dev:test": "vitest watch --ui",
 		"dev:pkg": "tsup --watch",
+		"dev:test": "vitest watch --ui",
 		"fix": "concurrently \"pnpm:fix:*\" --group --timings",
 		"fix:format": "pretty-quick --branch main --pattern \"packages/package/**/*\" --verbose",
 		"fix:js": "pnpm lint:js --fix",
@@ -61,16 +77,5 @@
 		"lint:pkg": "syncpack list-mismatches",
 		"lint:types": "tsc --noEmit",
 		"test:unit": "vitest run --dir \"./source\""
-	},
-	"peerDependencies": {
-		"typescript": "5.0.2"
-	},
-	"devDependencies": {
-		"@terminal-nerds/snippets-test": "workspace:*"
-	},
-	"peerDependenciesMeta": {
-		"typescript": {
-			"optional": true
-		}
 	}
 }
