--- packages/package/source/json/json.test.ts
@@ -1,10 +1,0 @@
-import { returns, throws } from "@terminal-nerds/snippets-test/unit";
-import { describe, it } from "vitest";
-
-import {} from "./json.ts";
-
-describe(``, () => {
-	it(``, ({ expect }) => {
-		expect();
-	});
-});
