+++ packages/package/source/main.ts
@@ -1,2 +1,4 @@
 /* MODULES */
+export * from "./dependencies/dependencies.ts";
 export * from "./json/json.ts";
+export * from "./schema/schema.ts";
